+++ src/cmd/make/doname.c
@@ -175,7 +175,8 @@ int doname(struct nameblock *p, int reclevel, TIMETYPE *tval)
         for (q = lp->depp; q; q = q->nxtdepblock) {
             errstat += doname(q->depname, reclevel + 1, &td1);
             if (dbgflag)
-                printf("TIME(%s)=%ld\n", q->depname->namep, td1);
+                printf("TIME(%s)=%lld\n", q->depname->namep,
+                    (long long)td1);
             if (td1 > td)
                 td = td1;
             if (ptime < td1)
@@ -234,7 +235,8 @@ int doname(struct nameblock *p, int reclevel, TIMETYPE *tval)
                             if (ptime < td)
                                 qchain = appendq(qchain, p2->namep);
                             if (dbgflag)
-                                printf("TIME(%s)=%ld\n", p2->namep, td);
+                                printf("TIME(%s)=%lld\n", p2->namep,
+                                    (long long)td);
                             if (td > tdep)
                                 tdep = td;
                             setvar("*", prefix);
