+++ src/cmd/uname/uname.c
@@ -99,7 +99,7 @@ main(
 		len = sizeof(buf);
 		if (sysctl(mib, 2, &buf, &len, NULL, 0) == -1)
 			err(1, "sysctl");
-		(void)printf("%s%.*s", prefix, len, buf);
+		(void)printf("%s%.*s", prefix, (int)len, buf);
 		prefix = " ";
 	}
 	if (flags & NFLAG) {
@@ -108,7 +108,7 @@ main(
 		len = sizeof(buf);
 		if (sysctl(mib, 2, &buf, &len, NULL, 0) == -1)
 			err(1, "sysctl");
-		(void)printf("%s%.*s", prefix, len, buf);
+		(void)printf("%s%.*s", prefix, (int)len, buf);
 		prefix = " ";
 	}
 	if (flags & RFLAG) {
@@ -117,7 +117,7 @@ main(
 		len = sizeof(buf);
 		if (sysctl(mib, 2, &buf, &len, NULL, 0) == -1)
 			err(1, "sysctl");
-		(void)printf("%s%.*s", prefix, len, buf);
+		(void)printf("%s%.*s", prefix, (int)len, buf);
 		prefix = " ";
 	}
 	if (flags & VFLAG) {
@@ -129,7 +129,7 @@ main(
 		for (p = buf, tlen = len; tlen--; ++p)
 			if (*p == '\n' || *p == '\t')
 				*p = ' ';
-		(void)printf("%s%.*s", prefix, len, buf);
+		(void)printf("%s%.*s", prefix, (int)len, buf);
 		prefix = " ";
 	}
 	if (flags & MFLAG) {
@@ -138,7 +138,7 @@ main(
 		len = sizeof(buf);
 		if (sysctl(mib, 2, &buf, &len, NULL, 0) == -1)
 			err(1, "sysctl");
-		(void)printf("%s%.*s", prefix, len, buf);
+		(void)printf("%s%.*s", prefix, (int)len, buf);
 		prefix = " ";
 	}
 	(void)printf("\n");
